Use the post ID in the permalink if unpublished
WordPress doesn’t create a “post_name” until a post is published, so
the permalink structure doesn’t work if it relies on the post name. Use
the post ID if the post_name does not exist. From the query end, we
check to see if ?preview=true is set and use the id if the slug is
numeric.