Update Advocacy Targets to Big Bets
Add code to handle new terms in the advocacy targets taxonomy.
Major template changes to the archive views for advocacy target terms
as well as to the policies and resources archives.
Includes new short templates to display brief view in archives.
New sprite sheet for terms as well as new imagery.